backend/requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
transformers[torch]>=4.21.0
requests>=2.28.0
python-multipart>=0.0.5
torch>=2.2.4
keybert>=0.7.0
sentence-transformers>=2.2.2
nltk>=3.8.1
